+++ src/main/java/net/geforcemods/securitycraft/screen/SCManualScreen.java
@@ -268,6 +268,19 @@ public class SCManualScreen extends Screen {
 			return true;
 		}
 
+		if (Screen.hasControlDown() && subpages.size() > 1) {
+			switch ((int) Math.signum(scroll)) {
+				case -1:
+					nextSubpage();
+					break;
+				case 1:
+					previousSubpage();
+					break;
+			}
+
+			return true;
+		}
+
 		switch ((int) Math.signum(scroll)) {
 			case -1:
 				nextPage();